Remove macFUSE, bindfs, and FUSE dependencies
Delete the v0 legacy packages (internal/deps, internal/mount,
internal/state) and the hidden v0 CLI commands that were the only
consumers of macFUSE and bindfs. Update all documentation to remove
references to these dependencies. Mark the v0 spec as deprecated.
